Composite with uncertainty

Composite with uncertainty

compositeEnsembles(
  fTS,
  binvec,
  spread = TRUE,
  stanFun = standardizeMeanIteratively,
  ageVar = "age",
  gaussianizeInput = FALSE,
  alignInterpDirection = TRUE,
  scope = "climate",
  binFun = sampleEnsembleThenBinTs,
  ...
)

Arguments

fTS

TS object of the timeseries that you want to composite

binvec

A vector of bin edges for binning

spread

An option to spread values over adjacent empty years (if TRUE, spreadPaleoData() is used by binFun)

stanFun

Function to use for standardization

ageVar

Name of the time vector in fTS

gaussianizeInput

Optionally gaussianize input before compositing

alignInterpDirection

Optionally invert timeseries with negative interpretation_direction

scope

interpretation_scope for alignInterpDirection (default = 'climate')

binFun

Function to use for binning

...

parameters to pass to stanFun

Value

a list. The main composite is in the list$composite matrix (needs a better definition here) TODO