create paleoclimate composite ensembles from a list of LiPD timeseries. A revised version of compositeEnsembles by Chris Hancock (2024)

create paleoclimate composite ensembles from a list of LiPD timeseries. A revised version of compositeEnsembles by Chris Hancock (2024)

compositeEnsembles2(
  fTS,
  binvec,
  nens = 100,
  stanFun = standardizeOverRandomInterval,
  ageVar = "age",
  uncVar = NA,
  weights = NA,
  samplePct = 1,
  scale = FALSE,
  spread = TRUE,
  gaussianizeInput = FALSE,
  alignInterpDirection = TRUE,
  scope = "climate",
  binFun = sampleEnsembleThenBinTs,
  ...
)

Arguments

fTS

TS list of the timeseries that you want to composite

binvec

vector of time boundaries over which to bin

nens

the number of ensembles to create

stanFun

function to use for standardization (either standardizeOverRandomInterval, standardizeMeanIteratively, or standardizeOverInterval)

ageVar

specify the name the time variable (typically 'age' or 'year')

uncVar

uncertainty variable
- Use a string to identify a fTS variable name or use a number to provide a uniform value to all records.
- This is used if stanFun == sampleEnsembleThenBinTs but not for simpleBinTs.

weights

weights for calculating the composite mean.
- For example, you may want to weight each record in a region relative to the number of total records from a single site.
- Use a string to identify a fTS variable name, otherwise all values will be weighted equally.

samplePct

the percent of records to used for each ensemble.
- If length(fTS)*samplePct <= 3, this is ignored.
- Default = 1 which means that 100 percent of records are used for each ensemble

scale

TODO

spread

should values be interpolated between bins? (TRUE/FALSE)

gaussianizeInput

Force values to gaussian distribution before analysis (TRUE/FALSE)

alignInterpDirection

multiply values by -1 if scope_interpDirection == negative (TRUE/FALSE)

scope

the scope of the project (typically "climate" (default) or "isotope")

binFun

function to use for binning (either sampleEnsembleThenBinTs or simpleBinTS)

...

Arguments passed on to standardizeOverRandomInterval

searchRange

the age range within which to search for an interval

duration

length of the interval to standardize the data within. Must be <= than searchRange

minN

minimum number length of de within the interval to calculate

normalizeVariance

Should the pdm be scaled to a uniform mean (FALSE) or variance and mean (TRUE) (default = TRUE)

Value

a composite list class with three tibbles:
(1) The composite [dims = length(binvec) x (nens + 1 for age column)]
(2) the standardized values [dims = length(binvec) x (length(fTS) + 1 for age column)], and
(3) a count of when proxy values are used in the compilation [same dims as 2]