A function to plot the output of compositeEnsembles2

A function to plot the output of compositeEnsembles2

plotComposite(
  compositeList,
  ageUnits = "yr BP",
  valUnits = "standardized",
  title = "composite proxy record",
  combine = TRUE,
  ...
)

Arguments

compositeList

The output of compositeEnsembles2()

ageUnits

units for the age variable

valUnits

units for the composite (default = standardized)

title

a title for the combined plot (not used if combine == FALSE)

combine

output a combined figure using egg::ggarrange or provide a list of subplots

...

Arguments passed on to geoChronR::plotTimeseriesEnsRibbons

probs

a vector of probabilities to plot as ribbons. It will create bands as ribbons of quantiles moving inward. If there's an odd number, it plots the middle quantile as a line.

color.low

Color of the outermost band; the extreme quantiles of the distribution

color.high

Color of the innermost band; the central quantiles of the distribution

color.line

Line color (following ggplot rules)

color.vector

A vector (of length equal to the number of bands) that specifies the colors for the ribbons from the outermost band in (default = NA). Colors specified as string according to ggplot2 conventions. If present, this overrules color.high and color.low

line.width

Width of the line

Value

a combined or list of gg plots